Move agent initialization logic into separate module
