+++ projector-agent-initialization/build.gradle.kts
@@ -1,0 +1,20 @@
+plugins {
+  kotlin("jvm")
+}
+
+kotlin {
+  explicitApi()
+}
+
+repositories {
+  mavenCentral()
+}
+
+dependencies {
+  testImplementation("org.junit.jupiter:junit-jupiter-api:5.6.0")
+  testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine")
+}
+
+tasks.getByName<Test>("test") {
+  useJUnitPlatform()
+}
