+++ projector-server-core/build.gradle.kts
@@ -125,6 +125,7 @@ val integrationTest = task<Test>("integrationTest") {
 
 dependencies {
   api(project(":projector-common"))
+  implementation(project(":projector-agent-initialization"))
   implementation(project(":projector-util-agent"))
   implementation(project(":projector-util-logging"))
   implementation("org.javassist:javassist:$javassistVersion")
